Add subfiles for sepearate upload routes
